Add support for 64 bits unsigned literals.

Close the pull request #129

+++ eclipse/plugins/net.sf.orcc.cal/src/net/sf/orcc/cal/services/Typer.java
@@ -252,7 +252,7 @@ public class Typer extends CalSwitch<Type> {
 
 	@Override
 	public Type caseExpressionInteger(ExpressionInteger expression) {
-		BigInteger value = BigInteger.valueOf(expression.getValue());
+		BigInteger value = expression.getValue();
 		return IrFactory.eINSTANCE.createTypeIntOrUint(value);
 	}
